Fixed oversights

+++ Arduino/testTracking/testTracking.ino
@@ -13,14 +13,14 @@ const int colorLeft = A0;  // Left color sensor (analog pin)
 const int colorRight = A1; // Right color sensor (analog pin)
 
 // Motor speed settings (0-255 PWM values)
-const float forwardSpeed = 50;  // Normal forward movement speed
-const float backwardSpeed = 25; // Backward movement speed
-const float turnSpeed = 20;     // Turning speed
-const float sprintSpeed = 100;  // High speed when enemy is detected
+const int forwardSpeed = 50;  // Normal forward movement speed
+const int backwardSpeed = 25; // Backward movement speed
+const int turnSpeed = 20;     // Turning speed
+const int sprintSpeed = 100;  // High speed when enemy is detected
 
 // Sensor threshold values
-const float whiteThreshold = 200;    // Analog value threshold for detecting white surface
-const float distanceThreshold = 100; // Digital threshold for IR obstacle detection
+const int whiteThreshold = 200;    // Analog value threshold for detecting white surface
+const int distanceThreshold = 100; // Digital threshold for IR obstacle detection
 
 // Behavior control flags
 const bool hardTurn = true;                  // Enable hard turns (both motors in opposite directions)
@@ -30,8 +30,8 @@ const bool turnBackwardsWhenSeeWhite = true; // Back up before turning when whit
 const bool useFastPinWrites = true;
 
 // Timing values for movements (in milliseconds)
-const float turnTimeOnWhite = 10;     // Time to turn when white line detected
-const float backwardTimeOnWhite = 10; // Time to move backward when white line detected
+const int turnTimeOnWhite = 10;     // Time to turn when white line detected
+const int backwardTimeOnWhite = 10; // Time to move backward when white line detected
 
 // Initialize motor driver objects using PWM_DIR mode
 // PWM_DIR mode uses separate pins for speed (PWM) and direction
